Foundation for browse API
