+++ src/BrowseParameters.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Loupe\Loupe;
+
+use Loupe\Loupe\Internal\Search\AbstractQueryParameters;
+
+final class BrowseParameters extends AbstractQueryParameters
+{
+}
