+++ src/Loupe.php
@@ -38,6 +38,12 @@ final class Loupe
         return $this->engine->addDocuments($documents);
     }
 
+    public function browse(BrowseParameters $parameters): SearchResult
+    {
+        StaticCache::enterContext($this);
+        return $this->engine->browse($parameters);
+    }
+
     public function countDocuments(): int
     {
         StaticCache::enterContext($this);
